pyproject.toml
[project]
name = "valence"
version = "1.0.1"
description = "Personal knowledge substrate for AI agents - beliefs, conversations, patterns"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Chris Jacobs"}
]
keywords = [
    "knowledge-base",
    "ai",
    "claude",
    "mcp",
    "agent",
    "memory",
]

dependencies = [
    # Orobobos bricks
    "oro-db>=0.1.0",
    "oro-models>=0.1.0",
    "oro-confidence>=0.1.0",
    "oro-crypto>=0.1.0",
    "oro-identity>=0.1.0",
    "oro-consensus>=0.1.0",
    "oro-storage>=0.1.0",
    "oro-mcp-base>=0.1.0",
    "oro-compliance>=0.1.0",
    "oro-network>=0.1.0",
    "oro-embeddings>=0.1.0",
    "oro-privacy>=0.1.0",
    "oro-federation>=0.1.0",
    # Existing deps
    "mcp>=1.0",
    "openai>=1.0",
    "psycopg2-binary>=2.9",
    "asyncpg>=0.29.0",
    "numpy>=1.24",
    "pgvector>=0.2",
    "sentence-transformers>=2.2.0",
    "starlette>=0.40.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.0.0",
    "PyJWT>=2.11.0",
    "PyYAML>=6.0",
    "aiohttp>=3.9.0",
    "dnspython>=2.4.0",
    "cryptography>=42.0",
]

[project.optional-dependencies]
all = []
redis = [
    "redis>=4.5",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "pytest-timeout>=2.2",
    "ruff>=0.1",
    "mypy>=1.0",
    "black>=24.0",
    "types-PyYAML>=6.0",
    "requests>=2.31",
]

[project.scripts]
valence = "valence.cli:main"
valence-substrate = "valence.substrate.mcp_server:run"
valence-vkb = "valence.vkb.mcp_server:run"
valence-server = "valence.server.app:run"
valence-token = "valence.server.cli:main"
valence-federation = "valence.cli.federation:main"
valence-seed = "valence.cli.seed:main"
valence-router = "valence.cli.router:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/valence"]

[tool.ruff]
line-length = 150
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = [
    "E501",  # line-too-long - allow long SQL strings and descriptions
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (requires PostgreSQL database)",
    "requires_postgres: Tests that require a real PostgreSQL database",
    "slow: Tests that take a long time to run",
]
addopts = "-q --tb=line --no-header"
# For verbose local debugging: pytest -v --tb=short

[tool.coverage.run]
source = ["src/valence"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_ignores = false
disallow_untyped_defs = false
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "psycopg2",
    "psycopg2.*",
    "asyncpg",
    "asyncpg.*",
]
ignore_missing_imports = true
